+++ src/Components/Authenticate/Authenticate.jsx
@@ -4,11 +4,10 @@ import './Authenticate.css'
 function Authenticate() {
     return (
       <section>
-        <h1>Welcome to the Authentication component</h1> 
+        <h1>Sign In</h1>
         <AuthenticateSignUp />
       </section>
     )
 }
-// in the future, these classes would not have garbage values 
 
 export default Authenticate