+++ src/Components/Components.jsx
@@ -1,5 +1,5 @@
 import Home from './Home/Home.jsx'
-import Register from './Register/Register.jsx'
+import Courses from './Courses/Courses.jsx'
 import Authenticate from './Authenticate/Authenticate.jsx'
 import Navigation from './Navigation/Navigation.jsx'
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom'
@@ -10,7 +10,7 @@ function Components() {
           <Navigation />
           <Routes>
             <Route path="/" element={<Home />} />
-            <Route path="/register" element={<Register />} />
+            <Route path="/courses" element={<Courses />} />
             <Route path="/authenticate" element={<Authenticate />} />
           </Routes>
         </Router>