+++ src/Components/Courses/Courses.jsx
@@ -1,0 +1,13 @@
+import CoursesDropdown from './CoursesDropdown.jsx'
+import './Courses.css'
+
+function Courses() {
+  return (
+    <section>
+      <h1>Courses</h1>
+      <CoursesDropdown />
+    </section>    
+  )
+}
+
+export default Courses