+++ CHANGELOG.md
@@ -1,12 +1,19 @@
 # Change Log 
 ## v0.2.0
-- improved UI/UX with CSS on individual pages 
+added: 
 - added Authenticate, Navigation, and Register components 
 - added ideas for future work
 - added user interactivity and logging in back4app
+changed: 
+- changed from ND class courses to Jiu-Jitsu courses
+fixed: 
+- improved UI/UX with CSS on individual pages 
 
 ## v0.1.0
+added: 
 - set up project structure with React
 - created initial page (Home)
 - implemented basic styling from feature 2
 - defined basic dependencies
+changed:
+fixed: